use fork of tinyexpr to make binary size smaller, use floats instead of doubles. add tinyexpr to project file so includes can be resolved. Don't download libs if already available

+++ BulbChipSTM32C071FBPx/Core/Inc/model/mode_state.h
@@ -16,7 +16,7 @@ typedef struct {
     uint8_t currentSectionIndex;
     uint32_t sectionElapsedMs;
     te_expr *compiledExprs[3];
-    double t_var;
+    float t_var;
 } EquationChannelState;
 
 typedef struct {
